okkkk

+++ MinecraftMediaLibrary/src/main/java/com/github/pulsebeat02/nms/PacketHandler.java
@@ -8,9 +8,9 @@ import java.util.UUID;
 
 public interface PacketHandler {
 
-    void display(UUID[] viewers, int map, int mapWidth, int mapHeight, ByteBuffer rgb, int videoWidth, int xOffset, int yOffset);
+    void display(UUID[] viewers, long map, int mapWidth, int mapHeight, ByteBuffer rgb, int videoWidth, int xOffset, int yOffset);
 
-    void display(UUID[] viewers, int map, int mapWidth, int mapHeight, ByteBuffer rgb, int videoWidth);
+    void display(UUID[] viewers, long map, int mapWidth, int mapHeight, ByteBuffer rgb, int videoWidth);
 
     void display(UUID[] viewers, Entity[] entities, int[] data, int width);
 
@@ -20,9 +20,9 @@ public interface PacketHandler {
 
     boolean isMapRegistered(int id);
 
-    void unregisterMap(int id);
+    void unregisterMap(long id);
 
-    void registerMap(int id);
+    void registerMap(long id);
 
     Object onPacketInterceptOut(Player viewer, Object packet);
